worky/requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.3
Django==2.0.7
djangorestframework==3.8.2
httpie==0.9.9
idna==2.7
psycopg2==2.7.5
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
